findbugs: reference comparison of const value instead of long value
findbugs reason:
RC: Suspicious reference comparison to constant (RC_REF_COMPARISON_BAD_PRACTICE)

+++ frontend/webadmin/modules/webadmin/src/main/java/org/ovirt/engine/ui/webadmin/widget/table/column/QuotaUtilizationStatusColumn.java
@@ -34,7 +34,7 @@ public class QuotaUtilizationStatusColumn<IVdcQueryable> extends WebAdminImageRe
         if (quota.getStorageSizeGB() == null) {
             return false;
         }
-        return quota.getStorageSizeGB() != QuotaStorage.UNLIMITED
+        return quota.getStorageSizeGB().longValue() != QuotaStorage.UNLIMITED.longValue()
                 && quota.getStorageSizeGB() < quota.getStorageSizeGBUsage();
     }
